fix: remove escrow adapter check
