+++ crates/tap-agent/src/tap/context/escrow.rs
@@ -16,19 +16,30 @@ impl From<eventuals::error::Closed> for AdapterError {
     }
 }
 
+// we don't need these checks anymore because there are being done before triggering
+// a rav request.
+//
+// In any case, we don't want to fail a receipt because of this.
+// The receipt is fine, just the escrow account that is not.
 #[async_trait]
 impl EscrowAdapterTrait for TapAgentContext {
     type AdapterError = AdapterError;
 
-    async fn get_available_escrow(&self, signer: Address) -> Result<u128, AdapterError> {
-        self.escrow_adapter.get_available_escrow(signer).await
+    async fn get_available_escrow(&self, _signer: Address) -> Result<u128, AdapterError> {
+        Ok(0)
     }
 
-    async fn subtract_escrow(&self, signer: Address, value: u128) -> Result<(), AdapterError> {
-        self.escrow_adapter.subtract_escrow(signer, value).await
+    async fn subtract_escrow(&self, _signer: Address, _value: u128) -> Result<(), AdapterError> {
+        Ok(())
     }
 
     async fn verify_signer(&self, signer: Address) -> Result<bool, Self::AdapterError> {
-        self.escrow_adapter.verify_signer(signer).await
+        let escrow_accounts = self.escrow_accounts.borrow();
+        let sender = escrow_accounts
+            .get_sender_for_signer(&signer)
+            .map_err(|_| AdapterError::ValidationError {
+                error: format!("Could not find the sender for the signer {}", signer),
+            })?;
+        Ok(sender == self.sender)
     }
 }
